[FE] Bug Fix - Betting error
Fixed error on placing bet, causing application to logout with error re: "non-numeric value"

+++ app/Http/Controllers/OrdersController.php
@@ -481,7 +481,7 @@ class OrdersController extends Controller
                 } else if ($last2 == 0) {
                     $actualStake = $ceil;
                 } else if ($last2 > 50) {
-                    $actualStake = substr($ceil, 0, -2) + 1;
+                    $actualStake = (int) substr($ceil, 0, -2) + 1;
                     $actualStake .= '00';
                 }
